input,button등 컴포넌트화
